Adding API call into showFreeTierLimitReached
