+++ src/utils/notifications.ts
@@ -5,6 +5,7 @@ import { trackRepoHandler, openSyncIgnoreHandler, disconnectRepoHandler, reconne
 import { UserState } from './user_utils';
 import { generateWebUrl } from './url_utils';
 import { authHandler } from '../handlers/user_commands';
+import { getCanAwailTrial } from './pricing_utils';
 
 
 export const showSignUpButtons = () => {
@@ -118,7 +119,8 @@ export const showSyncIgnoredRepo = (repoPath: string, parentRepoPath: string) =>
 		});
 };
 
-export const showFreeTierLimitReached = (repoPath: string, isNewPrivateRepo=false, canAvailTrial: boolean) => {
+export const showFreeTierLimitReached = async (repoPath: string, isNewPrivateRepo=false, accessToken="") => {
+	let canAvailTrial = await getCanAwailTrial(accessToken);
 	const msg = getUpgradePlanMsg(repoPath, isNewPrivateRepo);
 	const button = canAvailTrial ? NOTIFICATION_BUTTON.TRY_PRO_FOR_FREE : NOTIFICATION_BUTTON.UPGRADE_TO_PRO;
 	const pricingUrl = generateWebUrl(WebPaths.PRICING);
