+++ src/utils/pricing_utils.ts
@@ -100,8 +100,7 @@ export class PlanLimitsHandler {
 			if (this.repoId) return await this.run();
 			// This is "Connect Repo"
 			const isNewPrivateRepo = errorCode === ErrorCodes.PRIVATE_REPO_COUNT_LIMIT_REACHED;
-			const canAvailTrial = await getCanAwailTrial(this.accessToken);
-			showFreeTierLimitReached(this.repoPath, isNewPrivateRepo, canAvailTrial);
+			showFreeTierLimitReached(this.repoPath, isNewPrivateRepo, this.accessToken);
 			return true;
 		}
 		return false;
